fix resource providers card
