+++ packages/components/dashboard/resource-providers-card/resource-providers-card.scss
@@ -4,7 +4,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  padding: 1rem;
+  padding: 0.7rem;
 }
 
 .nb-resource-providers-card__row-status {
