+++ packages/components/dashboard/resource-providers-card/resource-providers-card.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { PrometheusResponse } from '@openshift-console/dynamic-plugin-sdk';
+import { useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk';
 import * as _ from 'lodash';
 import { useTranslation } from 'react-i18next';
 import {
@@ -11,80 +11,44 @@ import {
   GalleryItem,
   Skeleton,
 } from '@patternfly/react-core';
-import { MCG_MS_PROMETHEUS_URL } from '../../../constants';
-import { getMetric } from '../../../utils';
-import { useCustomPrometheusPoll } from '../../../utils/hooks/custom-prometheus-poll';
-import {
-  ResourceProvidersItem,
-  ProviderType,
-} from './resource-providers-card-item';
+import { NOOBAA_PROVIDER_MAP, NS_PROGRESS } from '../../../constants';
+import { NamespaceStoreKind } from '../../../types';
+import { nameSpaceStoreResource } from '../../resources';
 import './resource-providers-card.scss';
 
-const RESOURCE_PROVIDERS_QUERY = {
-  PROVIDERS_TYPES: ' NooBaa_cloud_types',
-  UNHEALTHY_PROVIDERS_TYPES: 'NooBaa_unhealthy_cloud_types',
-  RESOURCES_LINK_QUERY: 'NooBaa_system_links',
-};
-
-const getProviderType = (provider: ProviderPrometheusData): string =>
-  _.get(provider, 'metric.type', null);
-const getProviderCount = (provider: ProviderPrometheusData): number =>
-  Number(_.get(provider, 'value[1]', null));
-
-const filterProviders = (allProviders: ProviderType): string[] => {
-  return _.keys(allProviders).filter((provider) => allProviders[provider] > 0);
-};
-
-const createProvidersList = (data: PrometheusResponse): ProviderType => {
-  const providers = _.get(data, 'data.result', null);
-  const providersList: ProviderType = {};
-  if (_.isNil(providers)) return {};
-  providers.forEach((provider) => {
-    providersList[getProviderType(provider)] = getProviderCount(provider);
-  });
-  return providersList;
-};
+const ResourceProvidersItem: React.FC<ResourceProvidersRowProps> =
+  // eslint-disable-next-line react/display-name
+  React.memo(({ title, count }) => (
+    <div className="co-inventory-card__item">
+      <div
+        className="nb-resource-providers-card__row-title"
+        data-test="nb-resource-providers-card"
+      >{`${count} ${title}`}</div>
+    </div>
+  ));
 
 const ResourceProviders: React.FC<{}> = () => {
   const { t } = useTranslation();
 
-  const [providersTypesResult, providersTypesError, providersTypesLoading] =
-    useCustomPrometheusPoll({
-      query: RESOURCE_PROVIDERS_QUERY.PROVIDERS_TYPES,
-      endpoint: 'api/v1/query' as any,
-      basePath: MCG_MS_PROMETHEUS_URL,
-    });
-  const [
-    unhealthyProvidersTypesResult,
-    unhealthyProvidersTypesError,
-    unhealthyProvidersTypesLoading,
-  ] = useCustomPrometheusPoll({
-    query: RESOURCE_PROVIDERS_QUERY.UNHEALTHY_PROVIDERS_TYPES,
-    endpoint: 'api/v1/query' as any,
-    basePath: MCG_MS_PROMETHEUS_URL,
-  });
-  const [resourcesLinksResponse, resourcesLinksError, resourcesLinksLoading] =
-    useCustomPrometheusPoll({
-      query: RESOURCE_PROVIDERS_QUERY.RESOURCES_LINK_QUERY,
-      endpoint: 'api/v1/query' as any,
-      basePath: MCG_MS_PROMETHEUS_URL,
-    });
-  const error =
-    !!providersTypesError ||
-    !!unhealthyProvidersTypesError ||
-    !!resourcesLinksError;
+  const [dataResources, dataResourcesLoaded, dataResourcesError] =
+    useK8sWatchResource<NamespaceStoreKind[]>(nameSpaceStoreResource);
 
-  const noobaaResourcesLink = getMetric(resourcesLinksResponse, 'resources');
-
-  const allProviders = createProvidersList(providersTypesResult);
-  const unhealthyProviders = createProvidersList(unhealthyProvidersTypesResult);
-
-  const providerTypes = filterProviders(allProviders);
-  const hasProviders = !_.isEmpty(allProviders);
-  const isLoading =
-    providersTypesLoading ||
-    unhealthyProvidersTypesLoading ||
-    resourcesLinksLoading;
+  const allProvidersMap = React.useMemo(
+    () =>
+      dataResourcesLoaded && !dataResourcesError
+        ? dataResources?.reduce(
+            (drMap: AllProvidersMap, dataResource: NamespaceStoreKind) => {
+              const providerType =
+                NOOBAA_PROVIDER_MAP[dataResource?.spec?.type];
+              dataResource?.status?.phase === NS_PROGRESS.READY &&
+                (drMap[providerType] = drMap[providerType] + 1 || 1);
+              return drMap;
+            },
+            {} as AllProvidersMap
+          )
+        : ({} as AllProvidersMap),
+    [dataResources, dataResourcesLoaded, dataResourcesError]
+  );
 
   return (
     <Card>
@@ -92,24 +56,18 @@ const ResourceProviders: React.FC<{}> = () => {
         <CardTitle>{t('Resource Providers')}</CardTitle>
       </CardHeader>
       <CardBody>
-        {isLoading ? (
+        {!dataResourcesLoaded ? (
           <Skeleton />
-        ) : error || !hasProviders ? (
+        ) : dataResourcesError || _.isEmpty(allProvidersMap) ? (
           <div className="nb-resource-providers-card__not-available text-secondary">
             {t('Not available')}
           </div>
         ) : (
           <Gallery hasGutter>
-            {providerTypes.map((provider) => {
+            {_.map(allProvidersMap, (count, provider) => {
               return (
                 <GalleryItem key={provider}>
-                  <ResourceProvidersItem
-                    count={allProviders[provider]}
-                    key={provider}
-                    link={noobaaResourcesLink}
-                    title={provider}
-                    unhealthyProviders={unhealthyProviders}
-                  />
+                  <ResourceProvidersItem count={count} title={provider} />
                 </GalleryItem>
               );
             })}
@@ -120,9 +78,13 @@ const ResourceProviders: React.FC<{}> = () => {
   );
 };
 
-type ProviderPrometheusData = {
-  metric: { [key: string]: any };
-  value?: [number, string | number];
+type AllProvidersMap = {
+  string: number;
+};
+
+type ResourceProvidersRowProps = {
+  count: number;
+  title: string;
 };
 
 export const ResourceProvidersCard = ResourceProviders;
