+++ packages/constants/index.ts
@@ -1,5 +1,4 @@
 export * from './common';
 export * from './noobaa';
-export * from './providers';
 export * from './prometheus';
 export * from './paths';
