--- packages/constants/providers.ts
@@ -1,31 +1,0 @@
-export enum BC_PROVIDERS {
-  AWS = 'AWS S3',
-  S3 = 'S3 Compatible',
-  PVC = 'PVC',
-  GCP = 'Google Cloud Storage',
-  AZURE = 'Azure Blob',
-  IBM = 'IBM COS',
-  FILESYSTEM = 'Filesystem',
-}
-
-export const AWS_REGIONS = [
-  'us-east-1',
-  'us-east-2',
-  'us-west-1',
-  'us-west-2',
-  'ca-central-1',
-  'eu-central-1',
-  'eu-west-1',
-  'eu-west-2',
-  'eu-west-3',
-  'eu-north-1',
-  'ap-east-1',
-  'ap-northeast-1',
-  'ap-northeast-2',
-  'ap-northeast-3',
-  'ap-southeast-1',
-  'ap-southeast-2',
-  'ap-south-1',
-  'me-south-1',
-  'sa-east-1',
-];
