+++ packages/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './providers';
 export * from './noobaa-utils';
 export * from './common';
 export * from './metrics';
