--- packages/utils/providers.ts
@@ -1,45 +1,0 @@
-import * as _ from 'lodash';
-import { BC_PROVIDERS } from '../constants';
-import { SecretModel } from '../models';
-import { getAPIVersion } from './selectors/k8s';
-
-export const secretPayloadCreator = (
-  provider: string,
-  namespace: string,
-  secretName: string,
-  field1: string,
-  field2 = ''
-) => {
-  const payload = {
-    apiVersion: getAPIVersion(SecretModel),
-    kind: SecretModel.kind,
-    stringData: {},
-    metadata: {
-      name: secretName,
-      namespace,
-    },
-    type: 'Opaque',
-  };
-
-  switch (provider) {
-    case BC_PROVIDERS.AZURE:
-      payload.stringData = {
-        AccountName: field1,
-        AccountKey: field2,
-      };
-      break;
-    case BC_PROVIDERS.IBM:
-      payload.stringData = {
-        IBM_COS_ACCESS_KEY_ID: field1,
-        IBM_COS_SECRET_ACCESS_KEY: field2,
-      };
-      break;
-    default:
-      payload.stringData = {
-        AWS_ACCESS_KEY_ID: field1,
-        AWS_SECRET_ACCESS_KEY: field2,
-      };
-      break;
-  }
-  return payload;
-};
